lcd_driver/requirements.txt
RPi.GPIO==0.7.1
RPLCD==1.3.1
smbus2==0.5.0
